Switch to hash-map for underlying storage structure
Previously we used an array-map for the underlying storage for the
DataFrame.  However, using an array-map could cause performance
problems.  It has linear lookup time. If there were a DF with an
exceptionally large number of columns, we might have problems.

Switched to a hash-map and a vector for the underlying storage now.
The vector maintains the order of the columns.