Fix rename-key to handle (= old-key new-key)
Previous behavior was that if the new-key is the same as the old key, it
would change the key name (as expected) and then delete the old key.
This would result in a loss of data.

New behavior checks to see if the old-key and new-key are the same.
If so, it returns the hash-map unchanged.