Switch to Apache Commons CSV parser
Previously was using clojure.data.csv.  This is only acceptable for very
small CSV files.  Anything too large takes an unacceptable amount of
time (minutes) to complete.

I tried out the Apache Commons CSV code, and it works wonderfully
(about 20 seconds to read in 1,000,000 rows with 26 columns.)

Made the change to the Explojure base to use Apache Commons CSV now
instead of clojure.data.csv.

This may have ramifications in the future for porting Explojure to
a ClojureScript project.