Create column from scalar values
The following functions expect a column vector:
 - add-col
 - replace-col

I made it so that if a scalar (more correctly, a non-sequential) value
is passed, that value will be repeated to create a vector of appropriate
length for the dataframe in question.

Thus, we can now say (add-col df :new-colname 7) to add a column of 7's
to the dataframe df.